refactor: split sync/async vllm worker ([1/2] of refactor vllm worker) (#900)
