+++ nemo_rl/models/generation/vllm/config.py
@@ -1,0 +1,35 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from typing import Any, NotRequired, TypedDict
+
+from nemo_rl.models.generation.interfaces import GenerationConfig
+
+
+class VllmSpecificArgs(TypedDict):
+    tensor_parallel_size: int
+    pipeline_parallel_size: int
+    gpu_memory_utilization: float
+    max_model_len: int
+    # Additional arguments for vLLM inserted by nemo rl based on the context of when vllm is used
+    skip_tokenizer_init: bool
+    async_engine: bool
+    load_format: NotRequired[str]
+    precision: NotRequired[str]
+    enforce_eager: NotRequired[bool]
+
+
+class VllmConfig(GenerationConfig):
+    vllm_cfg: VllmSpecificArgs
+    vllm_kwargs: NotRequired[dict[str, Any]]
