+++ pyrefly.toml
@@ -85,7 +85,9 @@ project-includes = [
     "nemo_rl/models/dtensor/parallelize.py",
     "nemo_rl/models/generation/__init__.py",
     "nemo_rl/models/generation/interfaces.py",
-    "nemo_rl/models/generation/vllm_backend.py",
+    "nemo_rl/models/generation/vllm/__init__.py",
+    "nemo_rl/models/generation/vllm/config.py",
+    "nemo_rl/models/generation/vllm/vllm_backend.py",
     "nemo_rl/models/huggingface/__init__.py",
     "nemo_rl/models/megatron/__init__.py",
     "nemo_rl/models/megatron/community_import.py",
